Draft: feat: Replaces Java Bioformats with Rust-Bioformats (#25)
* chore: Ported do pure rust bioformats

* chore: Parallel image reader pool builder

* chore: Keep ui responsive

* chore: Improved load

* fix: performance improvement in image moving

* chore: Updated data type

* chore: Read pyrMID

* chore: Fixed rgb

* chore: Adapt readme

* chore: Fixed unit test

* chore: Reduced shipping size

* chore: Licensing fix

* chore: Added about slint

* chore: Updated deps

* chore: Removed unused packages

* chore: Small readme fix

* chore: Allow to build BSD only version

* chore: Rework about dialog

* chore: Moved to docs

* chore: fmt fix

+++ crates/gui/src/editor/viewport_cache.rs
@@ -36,7 +36,7 @@ pub struct ReadContext {
     pub res_idx: i32,
     pub image_w: usize,
     pub image_h: usize,
-    pub bit_depth: u8,
+    pub bit_depth: u16,
     pub _nr_color_channels: u8,
     pub viewport_width: f32,
     pub viewport_height: f32,
@@ -541,7 +541,7 @@ impl ViewportCache {
 
     /// Create a new image reader or use an existing if the image is the same
     ///
-    /// This creates an image reader (Java wrapper).
+    /// This creates an image reader.
 
     #[inline(always)]
     fn calc_draw_pos(
